make universal component ScheduleTable
