+++ src/components/ServiceStationDetails/ServiceStationDetails.jsx
@@ -1,6 +1,5 @@
 // import { useRef, useState } from "react";
 import ServiceStationDetailsBottom from "../ServiceStationDetailsBottom/ServiceStationDetailsBottom";
-import ServiceStationDetailsTop from "../ServiceStationDetailsTop/ServiceStationDetailsTop.jsx";
 import ServiceStationDetailsAccordion from "../ServiceStationDetailsAccordion/ServiceStationDetailsAccordion.jsx";
 import styles from "./ServiceStationDetails.module.css";
 import { useState } from "react";
